@@ -5,6 +5,8 @@
   "main": "main.ts",
   "type": "commonjs",
   "scripts": {
+    "lint": "eslint ./src/**",
+		"lint:fix": "eslint ./src/** --fix",
     "start": "node dist/main.js",
     "build": "tsc",
     "test": "echo \"Error: no test specified\" && exit 1"
@@ -16,10 +18,16 @@
     "express": "^4.17.2",
     "inversify": "^6.0.1",
     "reflect-metadata": "^0.1.13",
-    "tslog": "^3.3.2",
-    "typescript": "^4.5.4"
+    "tslog": "^3.3.2"
   },
   "devDependencies": {
-    "@types/express": "^4.17.13"
+    "@types/express": "^4.17.13",
+    "@typescript-eslint/eslint-plugin": "^5.30.7",
+    "@typescript-eslint/parser": "^5.30.7",
+    "eslint": "^8.20.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "prettier": "^2.7.1",
+    "typescript": "^4.7.4"
   }
 }
