@@ -1,49 +1,50 @@
-import express, { Express } from "express";
-import { Server } from "http";
-import { inject, injectable } from "inversify";
-import { ExeptionFilter } from "./errors/exeption.filter";
-import { ILogger } from "./logger/logger.interface";
-import { LoggerService } from "./logger/logger.service";
-import { TYPES } from "./types";
-import { UserController } from "./users/users.controller";
+import express, { Express } from 'express';
+import { Server } from 'http';
+import { inject, injectable } from 'inversify';
+import { ExeptionFilter } from './errors/exeption.filter';
+import { ILogger } from './logger/logger.interface';
+import { LoggerService } from './logger/logger.service';
+import { TYPES } from './types';
+import { UserController } from './users/users.controller';
 import 'reflect-metadata';
 
 @injectable()
 export class App {
-    app: Express;
-    server: Server;
-    port: number;
+  app: Express;
+  server: Server;
+  port: number;
 
-    constructor(
-        @inject(TYPES.ILogger) private logger: ILogger,
-        @inject(TYPES.UserController) private userController: UserController,
-        @inject(TYPES.ExeptionFilter) private exeptionFilter: ExeptionFilter,
-    ) {
-        this.app = express();
-        this.port = 8000;
-    }
+  constructor(
+    @inject(TYPES.ILogger) private logger: ILogger,
+    @inject(TYPES.UserController) private userController: UserController,
+    @inject(TYPES.ExeptionFilter) private exeptionFilter: ExeptionFilter,
+  ) {
+    this.app = express();
+    this.port = 8000;
+  }
 
-    useRoutes() {
-        this.app.use('/users', this.userController.router)
-    }
+  useRoutes(): void {
+    this.app.use('/users', this.userController.router);
+  }
 
-    useExeptionFilters() {
-        this.app.use(this.exeptionFilter.catch.bind(this.exeptionFilter));
-    }
+  useExeptionFilters(): void {
+    this.app.use(this.exeptionFilter.catch.bind(this.exeptionFilter));
+  }
 
-    public async init() {
-        this.useRoutes();
-        this.useExeptionFilters();
-        this.server = this.app.listen(this.port);
-        this.logger.log(`Сервер запущен на http://localhost:${this.port}`);
-    }
+  public async init(): Promise<void> {
+    this.useRoutes();
+    this.useExeptionFilters();
+    this.server = this.app.listen(this.port);
+    this.logger.log(`Сервер запущен на http://localhost:${this.port}`);
+  }
 }
 
-
+// пример самого простого ответа
 // app.get('/hello', (req, res) => {
 //     res.send('Привет!');
 // });
 
+// пример выброса необработанной ошибки
 // app.get('/error', (req, res) => {
 //     throw new Error('Всё пропало!!!!');
 // });
