@@ -1,40 +1,40 @@
 import { Router, Response } from 'express';
-import { IControllerRoute } from './route.interface';
+import { ExpressReturnType, IControllerRoute } from './route.interface';
 import { ILogger } from '../logger/logger.interface';
 import { injectable } from 'inversify';
 import 'reflect-metadata';
 
 @injectable()
 export abstract class BaseController {
-    private readonly _router: Router;
+  private readonly _router: Router;
 
-    constructor(private logger: ILogger) {
-        this._router = Router();
+  constructor(private logger: ILogger) {
+    this._router = Router();
+  }
+
+  get router(): Router {
+    return this._router;
+  }
+
+  public send<T>(res: Response, code: number, message: T): ExpressReturnType {
+    res.type('application/json');
+    return res.status(code).json(message);
+  }
+
+  public ok<T>(res: Response, message: T): ExpressReturnType {
+    return this.send<T>(res, 200, message);
+  }
+
+  public created(res: Response): ExpressReturnType {
+    return res.sendStatus(201);
+  }
+
+  protected bindRoutes(routes: IControllerRoute[]): void {
+    for (const route of routes) {
+      this.logger.log(`[${route.method}] ${route.path}`);
+      // чтобы не терять контекст
+      const handler = route.func.bind(this);
+      this.router[route.method](route.path, handler);
     }
-
-    get router() {
-        return this._router;
-    }
-
-    public send<T>(res: Response, code: number, message: T) {
-        res.type('application/json');
-        return res.status(code).json(message);
-    }
-
-    public ok<T>(res: Response, message: T) {
-        return this.send<T>(res, 200, message);
-    }    
-
-    public created(res: Response) {
-        return res.sendStatus(201);
-    }
-
-    protected bindRoutes(routes: IControllerRoute[]) {
-        for (const route of routes) {
-            this.logger.log(`[${route.method}] ${route.path}`);
-            // чтобы не терять контекст
-            const handler = route.func.bind(this);
-            this.router[route.method](route.path, handler);
-        }
-    }
+  }
 }
