@@ -1,7 +1,9 @@
-import { NextFunction, Request, Response, Router } from "express";
+import { NextFunction, Request, Response, Router } from 'express';
 
 export interface IControllerRoute {
-    path: string;
-    func: (req: Request, res: Response, next: NextFunction) => void;
-    method: keyof Pick<Router, 'get' | 'post' | 'delete' | 'patch' | 'put'>;
+  path: string;
+  func: (req: Request, res: Response, next: NextFunction) => void;
+  method: keyof Pick<Router, 'get' | 'post' | 'delete' | 'patch' | 'put'>;
 }
+
+export type ExpressReturnType = Response<any, Record<string, any>>;
