@@ -1,11 +1,11 @@
 export class HTTPError extends Error {
-    statusCode: number;
-    context?: string;
+  statusCode: number;
+  context?: string;
 
-    constructor(statusCode: number, message: string, context?: string) {
-        super(message)
-        this.statusCode = statusCode;
-        this.message = message;
-        this.context = context; 
-    }
+  constructor(statusCode: number, message: string, context?: string) {
+    super(message);
+    this.statusCode = statusCode;
+    this.message = message;
+    this.context = context;
+  }
 }
