@@ -1,8 +1,8 @@
 import { Logger } from 'tslog';
 
 export interface ILogger {
-    logger: Logger;
-    log: (...args: unknown[]) => void;
-    error: (...args: unknown[]) => void;
-    warn: (...args: unknown[]) => void;
+  logger: Logger;
+  log: (...args: unknown[]) => void;
+  error: (...args: unknown[]) => void;
+  warn: (...args: unknown[]) => void;
 }
