@@ -5,26 +5,26 @@ import 'reflect-metadata';
 
 @injectable()
 export class LoggerService implements ILogger {
-    public logger: Logger;
+  public logger: Logger;
 
-    constructor() {
-        this.logger = new Logger({
-            displayInstanceName: false,
-            displayLoggerName: false,
-            displayFilePath: 'hidden',
-            displayFunctionName: false,
-        });
-    }
+  constructor() {
+    this.logger = new Logger({
+      displayInstanceName: false,
+      displayLoggerName: false,
+      displayFilePath: 'hidden',
+      displayFunctionName: false,
+    });
+  }
 
-    log(...args: unknown[]) {
-        this.logger.info(...args);
-    }
+  log(...args: unknown[]): void {
+    this.logger.info(...args);
+  }
 
-    error(...args: unknown[]) {
-        this.logger.error(...args);
-    }
+  error(...args: unknown[]): void {
+    this.logger.error(...args);
+  }
 
-    warn(...args: unknown[]) {
-        this.logger.warn(...args);
-    }
+  warn(...args: unknown[]): void {
+    this.logger.warn(...args);
+  }
 }
